feat: add update chat, avatar url to chat, check participant rights, repos methods(get, update, can)

+++ internal/services/chatService/service.go
@@ -16,6 +16,9 @@ type ChatRepositoryInterface interface {
 	AddParticipant(participant entity.ChatParticipant) error
 	DeleteChat(chatID uint) error
 	DirectedChatCreated(firstId, secondId uint) (uint, error)
+	GetChatById(chatID uint) (*entity.Chat, error)
+	UpdateChat(chat *entity.Chat) (*entity.Chat, error)
+	UserCanChange(userID, chatID uint) (bool, error)
 }
 
 type ChatService struct {
@@ -115,7 +118,7 @@ func (s *ChatService) CreateChat(userID string, req request.CreateChatRequest) (
 		}
 	}
 
-	slog.Debug("creating chat completed")
+	slog.Debug("creating chat completed", "chat_id", createdChat.ID)
 	return createdChat.ID, nil
 }
 
@@ -126,17 +129,24 @@ func (s *ChatService) DeleteChat(userID string, req request.ChatRequest) error {
 		slog.Error("failed validate request", "error", err)
 		return err
 	}
-	// id, err := strconv.ParseUint(userID, 10, 32)
-	// if err != nil {
-	// 	slog.Error("failed parse user_id to uint", "user_id", userID)
-	// 	return errors.New("invalid user_id")
-	// }
-	// we need check user and his role for can he delete or not
+	id, err := strconv.ParseUint(userID, 10, 32)
+	if err != nil {
+		slog.Error("failed parse user_id to uint", "user_id", userID)
+		return errors.New("invalid user_id")
+	}
 
 	chatId, err := strconv.ParseUint(req.Id, 10, 32)
 	if err != nil {
-		slog.Error("failed parse chat_id to uint", "user_id", userID)
+		slog.Error("failed parse chat_id to uint", "chat_id", req.Id)
 		return errors.New("invalid chat_id")
+	}
+	can, err := s.repository.UserCanChange(uint(id), uint(chatId))
+	if err != nil {
+		slog.Error("failed get participant info", "error", err)
+		return err
+	}
+	if !can {
+		return errors.New("participant doesn't have permission to delete this chat")
 	}
 
 	err = s.repository.DeleteChat(uint(chatId))
@@ -145,6 +155,68 @@ func (s *ChatService) DeleteChat(userID string, req request.ChatRequest) error {
 		return err
 	}
 
-	slog.Debug("deleting chat completed")
+	slog.Debug("deleting chat completed", "chat_id", chatId)
 	return nil
+}
+
+func (s *ChatService) UpdateChat(userID string, req request.UpdateChatRequest) (*entity.Chat, error) {
+	slog.Debug("chat update started")
+
+	err := req.Validate()
+	if err != nil {
+		slog.Error("failed validate request", "error", err)
+		return nil, err
+	}
+
+	chatId, err := strconv.ParseUint(req.Id, 10, 32)
+	if err != nil {
+		slog.Error("failed parse chat_id to uint", "chat_id", req.Id)
+		return nil, errors.New("invalid chat_id")
+	}
+	id, err := strconv.ParseUint(userID, 10, 32)
+	if err != nil {
+		slog.Error("failed parse user_id to uint", "user_id", userID)
+		return nil, errors.New("invalid user_id")
+	}
+	can, err := s.repository.UserCanChange(uint(id), uint(chatId))
+	if err != nil {
+		slog.Error("failed get participant info", "error", err)
+		return nil, err
+	}
+	if !can {
+		return nil, errors.New("participant doesn't have permission to update this chat")
+	}
+
+	chat, err := s.repository.GetChatById(uint(chatId))
+	if err != nil {
+		slog.Error("failed get chat", "chat_id", req.Id)
+		return nil, err
+	}
+
+	// if chat directed not updated data
+	if chat.Type == entity.ChatTypeDirect {
+		slog.Error("cant update direct chat", "chat_id", chatId)
+		return nil, errors.New("failed update chat, direct chat not updated")
+	}
+
+	// update chat from req
+	if req.Name != "" {
+		chat.Name = req.Name
+	}
+	if req.Description != nil {
+		chat.Description = req.Description
+	}
+	if req.AvatarURL != nil {
+		chat.AvatarURL = req.AvatarURL
+	}
+	chat.IsPrivate = req.IsPrivate
+
+	updatedChat, err := s.repository.UpdateChat(chat)
+	if err != nil {
+		return nil, err
+	}
+
+	slog.Debug("chat updated successfully", "chat_id", chatId)
+	return updatedChat, nil
+
 }
